use unsigned 32bit integer for gpio pins

+++ include/uart_iqrf.h
@@ -65,12 +65,12 @@ typedef struct {
     /** Device file name*/
     char uartDev[UART_DEV_CAPACITY+1];
     int baudRate;
-    int8_t powerEnableGpioPin;      // GPIO to enable power supply to TR module
-    int8_t busEnableGpioPin;        // GPIO to enable function of SPI bus
-    int8_t pgmSwitchGpioPin;        // GPIO to switch TR module to PGM mode
-    int8_t spiEnableGpioPin;        // GPIO to enable function of SPI bus
-    int8_t uartEnableGpioPin;       // GPIO to enable function of UART bus
-    int8_t i2cEnableGpioPin;        // GPIO to enable function of I2C bus
+    int64_t powerEnableGpioPin;      // GPIO to enable power supply to TR module
+    int64_t busEnableGpioPin;        // GPIO to enable function of SPI bus
+    int64_t pgmSwitchGpioPin;        // GPIO to switch TR module to PGM mode
+    int64_t spiEnableGpioPin;        // GPIO to enable function of SPI bus
+    int64_t uartEnableGpioPin;       // GPIO to enable function of UART bus
+    int64_t i2cEnableGpioPin;        // GPIO to enable function of I2C bus
     tr_module_reset trModuleReset;  // enable / disable TR module reset during library initialization
 } T_UART_IQRF_CONFIG_STRUCT;
